Revert "Merge branch 'Development' into Feature-Music-Implementation"
This reverts commit 81aa62e8d5f7f195a86343faf3527aaa984bcd78, reversing
changes made to 99909ac09e6d4d95c1aac9aa66b9b17244b02479.

+++ Assets/Scripts/Organism/Bucket.cs
@@ -21,7 +21,7 @@ public class Bucket : MonoBehaviour
     private OrganismAnim organismToFollow;
     private int savedAmount;
     public int CurrentState => animateAlongSpline.CurrentState;
-    public Action<float> OnEmptyBucket;
+    public Action<int> OnEmptyBucket;
 
     private void OnValidate()
     {
